@@ -1,8 +1,11 @@
 package org.purplei2p.i2pd;
 
+import android.app.ActionBar;
 import android.app.Activity;
 import android.content.ComponentName;
+import android.content.Context;
 import android.content.Intent;
+import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
 import android.content.pm.ResolveInfo;
 import android.net.Uri;
@@ -17,6 +20,8 @@ import android.widget.CompoundButton;
 import android.widget.Switch;
 
 
+import org.purplei2p.i2pd.receivers.BootUpReceiver;
+
 import java.io.File;
 import java.util.List;
 import java.util.Objects;
@@ -25,18 +30,14 @@ import java.util.Objects;
 //import org.purplei2p.i2pd.iniedotr.IniEditor;
 
 public class SettingsActivity extends Activity {
-    private String TAG = "i2pdSrvcSettings";
-    private File cacheDir;
-    public static String onBootFileName = "/onBoot"; // just file, empty, if exist the do autostart, if not then no.
+    private final static String TAG = "Settings";
 
-    //https://gist.github.com/chandruark/3165a5ee3452f2b9ec7736cf1b4c5ea6
-    private void addAutoStartupSwitch() {
-
+    /** https://gist.github.com/chandruark/3165a5ee3452f2b9ec7736cf1b4c5ea6 */
+    private void maybeStartManufacturerSpecificBootupPermissionManagerActivity() {
         try {
             Intent intent = new Intent();
-            String manufacturer = android.os.Build.MANUFACTURER .toLowerCase();
-
-            switch (manufacturer){
+            String manufacturer = android.os.Build.MANUFACTURER.toLowerCase();
+            switch (manufacturer) {
                 case "xiaomi":
                     intent.setComponent(new ComponentName("com.miui.securitycenter", "com.miui.permcenter.autostart.AutoStartManagementActivity"));
                     break;
@@ -53,73 +54,45 @@ public class SettingsActivity extends Activity {
                     intent.setComponent(new ComponentName("com.huawei.systemmanager", "com.huawei.systemmanager.optimize.process.ProtectActivity"));
                     break;
                 case "oneplus":
-                    intent.setComponent(new ComponentName("com.oneplus.security", "com.oneplus.security.chainlaunch.view.ChainLaunchAppListAct‌​ivity"));
+                    intent.setComponent(new ComponentName("com.oneplus.security", "com.oneplus.security.chainlaunch.view.ChainLaunchAppListActivity"));
                     break;
             }
 
             List<ResolveInfo> list = getPackageManager().queryIntentActivities(intent, PackageManager.MATCH_DEFAULT_ONLY);
-            if (list.size() > 0) {
+            if (!list.isEmpty()) {
                 startActivity(intent);
             }
-        } catch (Exception e) {
-            Log.e("exceptionAutostarti2pd" , String.valueOf(e));
-        }
-
-    }
-
-    //@Override
-    private void requestPermission() {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            if (!Settings.canDrawOverlays(this)) {
-                Intent intent = new Intent(
-                        Settings.ACTION_MANAGE_OVERLAY_PERMISSION,
-                        Uri.parse("package:" + getPackageName())
-                );
-                startActivityForResult(intent, 232);
-            }
+        } catch (Throwable e) {
+            Log.e(TAG,"", e);
         }
     }
 
     public void onCreate(Bundle savedInstanceState) {
-        Log.d(TAG, "onCreate");
+        Log.d(TAG,"onCreate");
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_settings);
-        Objects.requireNonNull(getActionBar()).setDisplayHomeAsUpEnabled(true);
-        Switch autostart_switch = findViewById(R.id.autostart_enable);
-        Button openPreferences = findViewById(R.id.OpenPreferences);
-        cacheDir = getApplicationContext().getCacheDir();
-        File onBoot = new File(cacheDir.getAbsolutePath() + onBootFileName);
-        openPreferences.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent = new Intent(SettingsActivity.this, MainPreferenceActivity.class);
-                startActivity(intent);
+        final ActionBar actionBar = getActionBar();
+        if(actionBar!=null)actionBar.setDisplayHomeAsUpEnabled(true);
+        Switch autostartSwitch = findViewById(R.id.autostart_enable);
+        Button openPreferencesButton = findViewById(R.id.OpenPreferences);
+        File prefsDir = getApplicationContext().getFilesDir();
+        openPreferencesButton.setOnClickListener(view -> {
+            Intent intent = new Intent(SettingsActivity.this, MainPreferenceActivity.class);
+            startActivity(intent);
+        });
+        final SharedPreferences sharedPrefBootUp = getApplicationContext().getSharedPreferences(
+                BootUpReceiver.SHARED_PREF_FILE_KEY, Context.MODE_PRIVATE);
+        autostartSwitch.setOnCheckedChangeListener((view, isChecked) -> {
+            boolean autostartOnBootPrevValue = sharedPrefBootUp.getBoolean(BootUpReceiver.AUTOSTART_ON_BOOT, true);
+            SharedPreferences.Editor editor = sharedPrefBootUp.edit();
+            editor.putBoolean(BootUpReceiver.AUTOSTART_ON_BOOT, isChecked);
+            editor.apply();
+            if (isChecked && !autostartOnBootPrevValue) {
+                maybeStartManufacturerSpecificBootupPermissionManagerActivity();
             }
         });
-        autostart_switch.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-            public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-                // do something, the isChecked will be
-                // true if the switch is in the On position
-                if (isChecked) {
-                    if (!onBoot.exists()) {
-                        requestPermission();
-                        addAutoStartupSwitch();
-
-                        try {
-                            if (!onBoot.createNewFile())
-                                Log.d(TAG, "Cant create new wile on: "+onBoot.getAbsolutePath());
-                        } catch (Exception e) {
-                            Log.e(TAG, "error: " + e.toString());
-                        }
-                    }
-                } else {
-                    if (onBoot.exists())
-                        onBoot.delete();
-                }
-            }
-        });
-        if(onBoot.exists())
-            autostart_switch.setChecked(true);
+        boolean autostartOnBoot = sharedPrefBootUp.getBoolean(BootUpReceiver.AUTOSTART_ON_BOOT, true);
+        autostartSwitch.setChecked(autostartOnBoot);
     }
 
     public boolean onOptionsItemSelected(MenuItem item) {
