@@ -1,22 +1,42 @@
 package org.purplei2p.i2pd.appscope;
 
+import android.Manifest;
 import android.app.Application;
 import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
 import android.content.ServiceConnection;
+import android.content.SharedPreferences;
+import android.content.pm.PackageManager;
 import android.net.ConnectivityManager;
+import android.os.Build;
+import android.os.Environment;
 import android.os.IBinder;
 import android.util.Log;
+
+import androidx.core.content.ContextCompat;
 
 import org.purplei2p.i2pd.BuildConfig;
 import org.purplei2p.i2pd.I2PDActivity;
 import org.purplei2p.i2pd.*;
+import org.purplei2p.i2pd.receivers.BootUpReceiver;
+
+import java.lang.reflect.Method;
 
 public class App extends Application {
-    private static final String TAG = "i2pd.app";
+    private static final String TAG = "App";
 
-    //private static final I2PD_JNI jniHolder = new I2PD_JNI();
+    private static volatile boolean startDaemon = false;
+
+    public static synchronized boolean isStartDaemon() {
+        return startDaemon;
+    }
+
+    public static synchronized void setStartDaemon(boolean startDaemon) {
+        App.startDaemon = startDaemon;
+    }
+
+//private static final I2PD_JNI jniHolder = new I2PD_JNI();
 
     private static volatile DaemonWrapper daemonWrapper;
     private String versionName;
@@ -24,28 +44,62 @@ public class App extends Application {
     private static volatile boolean mIsBound;
 
 
-
     public synchronized static DaemonWrapper getDaemonWrapper() {
         return daemonWrapper;
     }
 
-    @Override
-    public void onCreate() {
-        super.onCreate();
-        synchronized (this) {
-            if (getDaemonWrapper() == null) {
-                createDaemonWrapper();
-            }
-            versionName = BuildConfig.VERSION_NAME;
-            doBindService();
-            startService(new Intent(this, ForegroundService.class));
+    public static void startForegroundService(Context context) {
+        Intent serviceIntent = new Intent(context, ForegroundService.class);
+        serviceIntent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            ContextCompat.startForegroundService(context, serviceIntent);
+        } else {
+            context.startService(serviceIntent);
         }
     }
 
-    private void createDaemonWrapper() {
-        ConnectivityManager  connectivityManager = (ConnectivityManager) getSystemService(
+    public static void maybeAutostartForegroundServiceOnBoot(Context context) {
+        boolean autostartOnBoot = isAutostartOnBoot(context);
+        if(autostartOnBoot) {
+            startForegroundService(context);
+        }
+    }
+
+    public static boolean isAutostartOnBoot(Context context) {
+        SharedPreferences sharedPref = context.getSharedPreferences(
+                BootUpReceiver.SHARED_PREF_FILE_KEY, MODE_PRIVATE);
+        boolean autostartOnBoot = sharedPref.getBoolean(BootUpReceiver.AUTOSTART_ON_BOOT, true);
+        return autostartOnBoot;
+    }
+
+    @Override
+    public void onCreate() {
+        Log.d(TAG, "App.onCreate");
+        super.onCreate();
+        if(App.isAutostartOnBoot(getApplicationContext())){
+            Log.d(TAG, "calling App.setStartDaemon(true)");
+            App.setStartDaemon(true);
+        }
+        if(App.isStartDaemon()){
+            Log.d(TAG, "calling App.doBindService()");
+            doBindService();
+        }
+        Log.d(TAG, "calling App.maybeAutostartForegroundServiceOnBoot()");
+        maybeAutostartForegroundServiceOnBoot(getApplicationContext());
+//        synchronized (this) {
+//            if (getDaemonWrapper() == null) {
+//                createDaemonWrapper();
+//            }
+        versionName = BuildConfig.VERSION_NAME;
+//            startService(new Intent(this, ForegroundService.class));
+//        }
+        Log.d(TAG, "App.onCreate() leave");
+    }
+
+    public void createDaemonWrapper() {
+        ConnectivityManager connectivityManager = (ConnectivityManager) getSystemService(
                 Context.CONNECTIVITY_SERVICE);
-        daemonWrapper = new DaemonWrapper(getApplicationContext(), getAssets(), connectivityManager);
+        daemonWrapper = new DaemonWrapper(this, getApplicationContext(), getAssets(), connectivityManager);
         ForegroundService.init(daemonWrapper);
     }
 
@@ -101,10 +155,11 @@ public class App extends Application {
 
     public synchronized void quit() {
         try {
-            if(daemonWrapper!=null)daemonWrapper.stopDaemon(null);
+            if (daemonWrapper != null) daemonWrapper.stopDaemon(null);
         } catch (Throwable tr) {
             Log.e(TAG, "", tr);
         }
+
         try {
             doUnbindService();
         } catch (IllegalArgumentException ex) {
@@ -115,12 +170,38 @@ public class App extends Application {
         } catch (Throwable tr) {
             Log.e(TAG, "throwable caught and ignored", tr);
         }
-        try{
+        try {
+            Log.d(TAG, "calling fgservice.stop");
             ForegroundService fs = ForegroundService.getInstance();
-            if(fs!=null)fs.stop();
-        }catch(Throwable tr) {
+            if (fs != null) fs.stop();
+        } catch (Throwable tr) {
             Log.e(TAG, "", tr);
         }
+    }
 
+    public boolean isPermittedToWriteToExternalStorage() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
+            return Environment.isExternalStorageManager();
+        } else {
+            Method methodCheckPermission;
+
+            try {
+                methodCheckPermission = getClass().getMethod(
+                        "checkSelfPermission", String.class);
+            } catch (NoSuchMethodException e) {
+                throw new RuntimeException(e);
+            }
+
+            Integer resultObj;
+
+            try {
+                resultObj = (Integer) methodCheckPermission.invoke(
+                        this, android.Manifest.permission.WRITE_EXTERNAL_STORAGE);
+            } catch (Throwable e) {
+                throw new RuntimeException(e);
+            }
+
+            return resultObj == PackageManager.PERMISSION_GRANTED;
+        }
     }
 }
